src/third_party/WebKit/LayoutTests/editing/shadow/delete-list-in-shadow-expected.txt

CONSOLE WARNING: line 22: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Elements distributed to Shadow DOM should be deleted correctly when pressing delete key.

To test manually, select somewhere in ABCDE from somehwere in 12345, and press delete, and check the selected text is deleted correctly.

PASS hostTextContent is 'AB345'
PASS successfullyParsed is true

TEST COMPLETE
